requirements.txt
Django==4.2.10
django-ninja==1.4.0
django-ninja-extra==0.22.9
django-ninja-jwt==5.3.7
django-simple-history==3.8.0
django-celery-results==2.6.0
pytest==7.4.3
pytest-django==4.7.0